Increased Modularity, fixed lang
-German translations, reduced lines in lang with universal fossil set numbers
-Made FossilBaseBodyBlock so Ocelot and Villager FossilBodies can inherit and save duplicate lines of code
