+++ src/main/java/net/Pandarix/betterarcheology/block/custom/VillagerFossilHeadBlock.java
@@ -7,7 +7,7 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.shape.VoxelShape;
 import net.minecraft.world.BlockView;
 
-public class VillagerFossilHeadBlock extends FossilHeadBlock{
+public class VillagerFossilHeadBlock extends FossilBaseHeadBlock {
     private static final VoxelShape VILLAGER_HEAD_SHAPE = Block.createCuboidShape(3, 0, 3, 13, 10, 13);
 
     public VillagerFossilHeadBlock(Settings settings) {
