+++ src/main/resources/assets/betterarcheology/lang/de_de.json
@@ -15,24 +15,23 @@
   "block.betterarcheology.cracked_mud_brick_stairs": "Rissige Schlammziegeltreppe",
   "block.betterarcheology.suspicious_red_sand": "Seltsamer Roter Sand",
   "block.betterarcheology.suspicious_dirt": "Seltsame Erde",
-  "block.betterarcheology.villager_fossil": "Dorfbewohner-Fossil",
-  "block.betterarcheology.villager_fossil_tooltip": "Hmmmm",
-  "block.betterarcheology.fossiliferous_dirt": "Fossilreiche Dirt",
-  "block.betterarcheology.fossiliferous_dirt_tooltip": "Vielleicht kann man es mit einer Schaufel ausgraben?",
-  "block.betterarcheology.villager_fossil_head": "Dorfbewohner-Fossil Schädel",
-  "block.betterarcheology.villager_fossil_head_tooltip": "Sieht nicht so aus, als hätten sie sich viel entwickelt...",
-  "block.betterarcheology.villager_fossil_head_count": " [1/2]",
-  "block.betterarcheology.villager_fossil_body": "Dorfbewohner-Fossil Körper",
-  "block.betterarcheology.villager_fossil_body_tooltip": "Sieht nicht so aus, als hätten sie sich viel entwickelt...",
-  "block.betterarcheology.villager_fossil_body_count": " [2/2]",
-  "block.betterarcheology.ocelot_fossil": "Ozelot-Fossil",
-  "block.betterarcheology.ocelot_fossil_tooltip": "Mreow",
+  "block.betterarcheology.fossiliferous_dirt": "Fossilreiche Erde",
+  "block.betterarcheology.fossiliferous_dirt_tooltip": "Vielleicht kann man was mit einer Schaufel ausgraben?",
+  "block.betterarcheology.villager_fossil": "Villager-Fossil",
+  "block.betterarcheology.villager_fossil_tooltip": "Die versteinerten Arme könnten als eine Art Aufsteller dienen.",
+  "block.betterarcheology.villager_fossil_head": "Villager-Fossil Skull",
+  "block.betterarcheology.villager_fossil_head_tooltip": "Sie scheinen sich nicht viel entwickelt zu haben...",
+  "block.betterarcheology.villager_fossil_body": "Villager-Fossil Body",
+  "block.betterarcheology.villager_fossil_body_tooltip": "Eine gruselige Erinnerung an ein Dorf, verschluckt von der Zeit",
+  "block.betterarcheology.ocelot_fossil": "Ocelot-Fossil",
+  "block.betterarcheology.ocelot_fossil_tooltip": "Creeper scheinen sie noch mehr zu fürchten als Katzen",
   "block.betterarcheology.ocelot_fossil_head": "Ocelot-Fossil Skull",
-  "block.betterarcheology.ocelot_fossil_head_tooltip": "They don't seem to have evolved a lot...",
-  "block.betterarcheology.ocelot_fossil_head_count": " [1/2]",
+  "block.betterarcheology.ocelot_fossil_head_tooltip": "Im Gegensatz zu ihren Nachkommen hatten sie lange und scharfe Zähne",
   "block.betterarcheology.ocelot_fossil_body": "Ocelot-Fossil Body",
-  "block.betterarcheology.ocelot_fossil_body_tooltip": "They don't seem to have evolved a lot...",
-  "block.betterarcheology.ocelot_fossil_body_count": " [2/2]",
+  "block.betterarcheology.ocelot_fossil_body_tooltip": "Es sieht weniger agil, aber stärker aus",
+
+  "block.betterarcheology.fossil_head_set": " [1/2]",
+  "block.betterarcheology.fossil_body_set": " [2/2]",
 
   "item.betterarcheology.iron_brush": "Eisener Pinsel",
   "item.betterarcheology.diamond_brush": "Diamantener Pinsel",
@@ -41,7 +40,7 @@
   "item.betterarcheology.unidentified_artifact": "Unidentifiziertes Artefakt",
   "item.betterarcheology.identified_artifact": "Identifiziertes Artefakt",
 
-  "entity.betterarcheology.villager.archeologist": "Archäologe",
+  "entity.minecraft.villager.archeologist": "Archäologe",
 
   "enchantment.betterarcheology.penetrating_strike": "Durchdringender Hieb",
   "item.betterarcheology.penetrating_strike_tooltip": "Ignoriert die Effekte von Schutz-Verzauberungen teilweise",
