+++ src/main/resources/assets/betterarcheology/lang/en_us.json
@@ -16,23 +16,22 @@
   "block.betterarcheology.suspicious_red_sand": "Suspicious Red Sand",
   "block.betterarcheology.suspicious_dirt": "Suspicious Dirt",
   "block.betterarcheology.fossiliferous_dirt": "Fossiliferous Dirt",
-  "block.betterarcheology.fossiliferous_dirt_tooltip": "Maybe it can be uncovered with a shovel?",
+  "block.betterarcheology.fossiliferous_dirt_tooltip": "Maybe something can be uncovered with a shovel?",
   "block.betterarcheology.villager_fossil": "Villager-Fossil",
   "block.betterarcheology.villager_fossil_tooltip": "Its fossilized arms could act as a sort of display",
   "block.betterarcheology.villager_fossil_head": "Villager-Fossil Skull",
   "block.betterarcheology.villager_fossil_head_tooltip": "They don't seem to have evolved a lot...",
-  "block.betterarcheology.villager_fossil_head_count": " [1/2]",
   "block.betterarcheology.villager_fossil_body": "Villager-Fossil Body",
   "block.betterarcheology.villager_fossil_body_tooltip": "A haunting reminder of a village lost to time",
-  "block.betterarcheology.villager_fossil_body_count": " [2/2]",
   "block.betterarcheology.ocelot_fossil": "Ocelot-Fossil",
   "block.betterarcheology.ocelot_fossil_tooltip": "Creepers seem to fear them even more than cats",
   "block.betterarcheology.ocelot_fossil_head": "Ocelot-Fossil Skull",
   "block.betterarcheology.ocelot_fossil_head_tooltip": "Unlike their descendants, they had long and sharp teeth",
-  "block.betterarcheology.ocelot_fossil_head_count": " [1/2]",
   "block.betterarcheology.ocelot_fossil_body": "Ocelot-Fossil Body",
   "block.betterarcheology.ocelot_fossil_body_tooltip": "It looks less agile but stronger",
-  "block.betterarcheology.ocelot_fossil_body_count": " [2/2]",
+
+  "block.betterarcheology.fossil_head_set": " [1/2]",
+  "block.betterarcheology.fossil_body_set": " [2/2]",
 
   "item.betterarcheology.iron_brush": "Iron Brush",
   "item.betterarcheology.diamond_brush": "Diamond Brush",
